requirements.txt
addict==2.4.0
certifi==2023.7.22
charset-normalizer==3.2.0
cmake==3.27.2
coloredlogs==15.0.1
contourpy==1.1.0
cycler==0.11.0
dill==0.3.7
easydict==1.10
filelock==3.12.3
flatbuffers==23.5.26
fonttools==4.42.1
humanfriendly==10.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.1
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
lit==16.0.6
MarkupSafe==2.1.3
matplotlib==3.7.1
mmcv==1.7.1
mpmath==1.3.0
multiprocess==0.70.14
networkx==3.1
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
onnxruntime-gpu==1.15.1
opencv-python==4.8.0.76
packaging==23.1
Pillow==10.0.0
platformdirs==3.10.0
prettytable==3.8.0
protobuf==4.24.2
pyparsing==3.1.1
python-dateutil==2.8.2
PyYAML==6.0.1
requests==2.31.0
scikit-learn==1.2.2
scipy==1.11.2
six==1.16.0
sympy==1.12
threadpoolctl==3.2.0
tomli==2.0.1
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
triton==2.0.0
typing_extensions==4.7.1
urllib3==2.0.4
wcwidth==0.2.6
yapf==0.40.1
zipp==3.16.2
